+++ src/i18n/index.js
@@ -1,3 +1,4 @@
 export { LanguageContext } from './LanguageContext';
 export { LanguageProvider } from './LanguageProvider';
 export { useTranslation } from './useTranslation';
+export { default as i18n } from './i18n';